1.移门清单、移门按门洞面积计算，黑色窄框/白色窄框/ 35框/65框 2.板件清单(圆弧踢脚板)部分的逻辑添加 3.添加上传附件的路径dir参数，好保存到数据库fullDir1， 4.item数量count改为decimal 5.添加报价编辑SaveMoneyForm保存 6.重复导入报价附件，先删掉之前导入的导入记录，不管是酷家乐或是1010 7.添加线条部分的逻辑:95顶线1#/75顶线1#按米计算      95顶线1#/75顶线1#长度=（宽度 >2400？长度=宽度/2000+0.2，数量=ceil（宽度/2400）：长度=宽度/1000+0.2） 8.展示1010，酷家乐导入状态 9.detail调整报价详情，.html()，没有删除列， form添加附件展示MoneyPath1010和MoneyPathKuJiaLe， style="width: 60px;"

+++ HZSoft.Application/HZSoft.Application.Entity/CustomerManage/DZ_Money_ItemEntity.cs
@@ -73,7 +73,7 @@ namespace HZSoft.Application.Entity.CustomerManage
         /// </summary>
         /// <returns></returns>
         [Column("COUNT")]
-        public int? Count { get; set; }
+        public decimal? Count { get; set; }
         /// <summary>
         /// Ãæ»ý
         /// </summary>
